+++ test/BaseApiTest.php
@@ -5,6 +5,7 @@ namespace OhMyBrew\Test;
 use GuzzleHttp\Psr7\Response;
 use GuzzleHttp\Psr7\Uri;
 use OhMyBrew\BasicShopifyAPI;
+use Psr\Log\NullLogger;
 use ReflectionClass;
 use ReflectionMethod;
 
@@ -34,6 +35,33 @@ class BaseApiTest extends BaseTest
 
         $this->assertEquals(false, $api->isPrivate());
         $this->assertEquals(true, $api->isPublic());
+    }
+
+    /**
+     * @test
+     *
+     * Should set a logger.
+     */
+    public function itShouldSetLogger()
+    {
+        $api = new BasicShopifyAPI();
+
+        // Confirm no logging works
+        $this->assertFalse($api->log('Hello world!'));
+
+        // Set the test logger
+        $api->setLogger(new NullLogger());
+
+        // Get the logger value
+        $reflected = new ReflectionClass($api);
+        $loggerProperty = $reflected->getProperty('logger');
+        $loggerProperty->setAccessible(true);
+
+        // Ensure logger value matches the test logger
+        $this->assertEquals(NullLogger::class, get_class($loggerProperty->getValue($api)));
+
+        // Confirm logging now works
+        $this->assertTrue($api->log('Hello world!'));
     }
 
     /**